fetcher/requirements.txt
Flask
pymongo
requests
schedule
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
python-dotenv
backoff